// types.ts
export interface CityWeather {
    coord: {
        lon: number;
        lat: number;
    };
    weather: {
        id: number;
        main: string;
        description: string;
        icon: string;
    }[];
    base: string;
    main: {
        temp: number;
        feels_like: number;
        temp_min: number;
        temp_max: number;
        pressure: number;
        humidity: number;
        sea_level: number;
        grnd_level: number;
    };
    visibility: number;
    wind: {
        speed: number;
        deg: number;
    };
    clouds: {
        all: number;
    };
    dt: number;
    sys: {
        type: number;
        id: number;
        country: string;
        sunrise: number;
        sunset: number;
    };
    timezone: number;
    id: number;
    name: string;
    cod: number;
}

export interface City {
    id: string;
    city: string;
    country: string;
    lat: number;
    lon: number;
    temperature: number;
}

export enum SortByOption {
    TEMPERATURE = 'temperature',
    CITY_ASC = 'city_asc',
    CONTINENT = 'continent',
}

export interface Coordinates {
    lon: number;
    lat: number;
}